Update all dependencies
Note: This shall be released as a major version for all updated packages.

Currently this cannot be merged as the `eslint-config-react-app` has not yet
released the changes to add support for eslint 8. Once they make the release
we can change `next` to the actual version number and also remove the resolution
for `eslint-plugin-flowtype`.